feat: search input value와 parameter 동기화하여 관리
