+++ src/main.ts
@@ -1,4 +1,5 @@
 import { getPopularMovies, Movie } from "./apis/movie/api";
+import { getSearchedMovies } from "./apis/search/api";
 import { handleMovieSearch } from "./dom/eventHandler/handleMovieSearch";
 import {
   handleMainSeeMore,
@@ -56,6 +57,28 @@ if (searchSeeMoreButton && searchInput) {
 // TODO: popular movies 없는 경우 배너 대체 UI
 
 const render = async () => {
+  renderInitialUI();
+
+  const url = new URL(window.location.href);
+  const params = url.searchParams;
+  const keyword = params.get("keyword");
+  if (keyword) {
+    await renderSearchUI(keyword);
+  } else {
+    await renderMainUI();
+  }
+};
+
+export const renderInitialUI = () => {
+  renderResultSectionContent({
+    isLoading: true,
+    isError: false,
+    isLastPage: false,
+    movies: [],
+  });
+};
+
+export const renderMainUI = async () => {
   let isError = false;
   let isLastPage = false;
   let movies: Movie[] = [];
@@ -79,12 +102,40 @@ const render = async () => {
   }
 };
 
-renderResultSectionContent({
-  isLoading: true,
-  isError: false,
-  isLastPage: false,
-  movies: [],
-});
+export const renderSearchUI = async (keyword: string) => {
+  const searchInput = document.getElementById(
+    "search-input",
+  ) as HTMLInputElement;
+  const banner = document.getElementById("background-container");
+  const resultSection = document.getElementById("result-section");
+  const subTitle = document.getElementById("sub-title");
+  const thumbnailListElement = document.getElementById("search-thumbnail-list");
+  searchInput.value = keyword;
+
+  if (!banner || !subTitle || searchInput?.value.trim() === "") return;
+
+  const searchResult = await getSearchedMovies({
+    query: keyword,
+    language: "ko-KR",
+    page: 1,
+  });
+  if (searchResult) {
+    banner.hidden = true;
+    resultSection?.classList.add("result-section");
+    subTitle.innerText = `"${keyword}" 검색 결과`;
+    const isLastPage = searchResult.page === searchResult.total_pages;
+    const movies = searchResult.results;
+    renderResultSectionContent({
+      isLoading: false,
+      isError: false,
+      isLastPage,
+      type: "search",
+      movies,
+    });
+    renderThumbnailList({ movies, thumbnailListElement });
+  }
+};
+
 await render();
 
 // 진입 -> renderResultSectionContent 호출 -> api 호출
